Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 // Your code here...
 #include <stdio.h>
 int main(){
-    int a, b, c;
-    scanf("%d %d %d", a, b, c);
+    float a, b, c;
+    scanf("%f %f %f", a, b, c);
     printf("Average: %.2f", (a+b+c)/2);
     return 0;
 }
